partial completion of karmavoucher import
