+++ src/App.tsx
@@ -13,7 +13,7 @@ import ResetPassword from "./modules/Common/Authentication/pages/ResetPassword";
 import PrivateRoutes from "./components/PrivateRoutes";
 import DashboardRootLayout from "./modules/Dashboard/layouts/DashboardRootLayout";
 import NotFound from "./components/NotFound";
-import Profile from "./modules/Dashboard/modules/Profile/pages/Profile"
+import Profile from "./modules/Dashboard/modules/Profile/pages/Profile";
 
 const Tasks = lazy(() =>
     import("./modules/Dashboard/modules/Tasks/Tasks").then(module => ({
@@ -170,6 +170,7 @@ import { CampusStudentList, ConnectDiscord } from "./modules/Dashboard/modules";
 import Refer from "./modules/Dashboard/modules/Refer/Refer";
 import Thread from "./modules/Public/ThreadsCard/modules/Thread";
 import { KarmaVoucher } from "./modules/Dashboard/modules/KarmaVoucher/KarmaVoucher";
+import KarmaVoucherBulkImport from "./modules/Dashboard/modules/KarmaVoucher/components/KarmaVoucherBulkImport";
 const ConnectedDevices = lazy(
     () => import("./modules/Dashboard/modules/Settings/pages/ConnectedDevices")
 );
@@ -223,7 +224,7 @@ function App() {
                         },
                         {
                             path: "refer",
-                            element: <Refer/>
+                            element: <Refer />
                         },
                         {
                             path: "interest-groups",
@@ -379,6 +380,10 @@ function App() {
                                     children={<KarmaVoucher />}
                                 />
                             )
+                        },
+                        {
+                            path: "karma-voucher/bulk-import",
+                            element: <KarmaVoucherBulkImport />
                         },
                         {
                             path: "url-shortener",
