+++ src/modules/Dashboard/modules/KarmaVoucher/KarmaVoucher.tsx
@@ -1,4 +1,7 @@
-import { MuButton, PowerfulButton } from "@/MuLearnComponents/MuButtons/MuButton";
+import {
+    MuButton,
+    PowerfulButton
+} from "@/MuLearnComponents/MuButtons/MuButton";
 import Pagination from "@/MuLearnComponents/Pagination/Pagination";
 import THead from "@/MuLearnComponents/Table/THead";
 import Table from "@/MuLearnComponents/Table/Table";
@@ -9,7 +12,7 @@ import { useEffect, useRef, useState } from "react";
 import { AiOutlinePlusCircle } from "react-icons/ai";
 import { useNavigate } from "react-router-dom";
 import styles from "../InterestGroup/InterestGroup.module.css";
-import { getKarmaVoucher } from "./KarmaVoucherApis";
+import { getKarmaVoucher } from "./service/api";
 
 type Props = {};
 
@@ -72,35 +75,73 @@ export const KarmaVoucher = (props: Props) => {
 
     useEffect(() => {
         if (firstFetch.current) {
-            getKarmaVoucher(setData, 1, perPage, setIsLoading, setTotalPages, "", "");
+            getKarmaVoucher(
+                setData,
+                1,
+                perPage,
+                setIsLoading,
+                setTotalPages,
+                "",
+                ""
+            );
         }
         firstFetch.current = false;
     }, [data]);
 
     const handleSearch = (search: string) => {
         setCurrentPage(1);
-        getKarmaVoucher(setData, 1, perPage, setIsLoading, setTotalPages, search, "");
+        getKarmaVoucher(
+            setData,
+            1,
+            perPage,
+            setIsLoading,
+            setTotalPages,
+            search,
+            ""
+        );
     };
 
     const handlePerPageNumber = (selectedValue: number) => {
         setCurrentPage(1);
         setPerPage(selectedValue);
-        getKarmaVoucher(setData, 1, selectedValue, setIsLoading, setTotalPages, "", "");
+        getKarmaVoucher(
+            setData,
+            1,
+            selectedValue,
+            setIsLoading,
+            setTotalPages,
+            "",
+            ""
+        );
     };
 
     const handleIconClick = (column: string) => {
         if (sort === column) {
             setSort(`-${column}`);
-            getKarmaVoucher(setData, 1, perPage, setIsLoading, setTotalPages, "", `-${column}`);
+            getKarmaVoucher(
+                setData,
+                1,
+                perPage,
+                setIsLoading,
+                setTotalPages,
+                "",
+                `-${column}`
+            );
         } else {
             setSort(column);
-            getKarmaVoucher(setData, 1, perPage, setIsLoading, setTotalPages, "", column);
+            getKarmaVoucher(
+                setData,
+                1,
+                perPage,
+                setIsLoading,
+                setTotalPages,
+                "",
+                column
+            );
         }
 
         //console.log(`Icon clicked for column: ${column}`);
     };
-
-   
 
     return (
         <>
@@ -110,8 +151,16 @@ export const KarmaVoucher = (props: Props) => {
                     gap: "15px"
                 }}
             >
-
-                
+                {/* Not Ready yet 
+                <PowerfulButton
+                    variant="secondary"
+                    onClick={() =>
+                        navigate("/dashboard/karma-voucher/bulk-import")
+                    }
+                >
+                    <AiOutlinePlusCircle />
+                    Bulk Import
+                </PowerfulButton> */}
             </div>
 
             {data && (
@@ -136,17 +185,19 @@ export const KarmaVoucher = (props: Props) => {
                             onIconClick={handleIconClick}
                         />
                         <div>
-                            {!isLoading &&
+                            {!isLoading && (
                                 <Pagination
                                     currentPage={currentPage}
                                     totalPages={totalPages}
                                     margin="10px 0"
                                     handleNextClick={handleNextClick}
-                                    handlePreviousClick={handlePreviousClick} onSearchText={handleSearch}
+                                    handlePreviousClick={handlePreviousClick}
+                                    onSearchText={handleSearch}
                                     onPerPageNumber={handlePerPageNumber}
                                     perPage={perPage}
                                     setPerPage={setPerPage}
-                                />}
+                                />
+                            )}
                         </div>
                         {/*use <Blank/> when u don't need <THead /> or <Pagination inside <Table/> cause <Table /> needs atleast 2 children*/}
                     </Table>
