+++ src/modules/Dashboard/modules/Tasks/TaskBulkImport.tsx
@@ -37,7 +37,7 @@ const TaskBulkImport = (props: Props) => {
         () => failureDownload,
         [uploadResponse]
     );
-
+    console.log(uploadResponse);
     return (
         <>
             <div
@@ -46,7 +46,6 @@ const TaskBulkImport = (props: Props) => {
                     justifyContent: "space-between",
                     width: "100%",
                     alignItems: "center"
-                
                 }}
             >
                 <SingleButton
@@ -56,7 +55,7 @@ const TaskBulkImport = (props: Props) => {
                         display: "flex",
                         justifyContent: "start",
                         width: "100%",
-                        alignItems: "center",
+                        alignItems: "center"
                     }}
                     onClick={() => {
                         navigate("/dashboard/tasks");
